requirements.txt
ollama
chromadb
tqdm
Pillow
